Move out file discovery methods to separate module
Methods responsible for the file discovery resided in Finder's module as
it was the only manager's component using them.  However, the backfill
tool will most likely be using one of them as well so I placed them in a
separate module.
